+++ Core/GameEngineDevice/Source/W3DDevice/Common/System/W3DRadar.cpp
@@ -197,8 +197,8 @@ void W3DRadar::reconstructViewBox( void )
 	{
 
 		// first convert to radar cells
- 		radar[ i ].x = world[ i ].x / (m_mapExtent.width() / RADAR_CELL_WIDTH);
- 		radar[ i ].y = world[ i ].y / (m_mapExtent.height() / RADAR_CELL_HEIGHT);
+ 		radar[ i ].x = world[ i ].x / (m_mapExtent.width() / static_cast<float>(RADAR_CELL_WIDTH));
+ 		radar[ i ].y = world[ i ].y / (m_mapExtent.height() / static_cast<float>(RADAR_CELL_HEIGHT));
 
 		//
 		// store these points in the view box array which contains a first position
@@ -237,9 +237,9 @@ void W3DRadar::radarToPixel( const ICoord2D *radar, ICoord2D *pixel,
 	if( radar == nullptr || pixel == nullptr )
 		return;
 
-	pixel->x = (radar->x * radarWidth / RADAR_CELL_WIDTH) + radarUpperLeftX;
+	pixel->x = (radar->x * radarWidth / static_cast<float>(RADAR_CELL_WIDTH)) + radarUpperLeftX;
 	// note the "inverted" y here to orient the way our world looks with +x=right and -y=down
-	pixel->y = ((RADAR_CELL_HEIGHT - 1 - radar->y) * radarHeight / RADAR_CELL_HEIGHT) + radarUpperLeftY;
+	pixel->y = ((static_cast<float>(RADAR_CELL_HEIGHT) - 1 - radar->y) * radarHeight / static_cast<float>(RADAR_CELL_HEIGHT)) + radarUpperLeftY;
 
 }
 
@@ -255,8 +255,8 @@ void W3DRadar::drawHeroIcon( Int pixelX, Int pixelY, Int width, Int height, cons
 	{
 		// convert world to radar coords
 		ICoord2D ulRadar;
-		ulRadar.x = pos->x / (m_mapExtent.width() / RADAR_CELL_WIDTH);
-		ulRadar.y = pos->y / (m_mapExtent.height() / RADAR_CELL_HEIGHT);
+		ulRadar.x = pos->x / (m_mapExtent.width() / static_cast<float>(RADAR_CELL_WIDTH));
+		ulRadar.y = pos->y / (m_mapExtent.height() / static_cast<float>(RADAR_CELL_HEIGHT));
 
 		// convert radar to screen coords
 		ICoord2D offsetScreen;
@@ -307,8 +307,8 @@ void W3DRadar::drawViewBox( Int pixelX, Int pixelY, Int width, Int height )
 	TheTacticalView->screenToWorldAtZ( &ulScreen, &ulWorld, getTerrainAverageZ() );
 
 	// convert world to radar coords
- 	ulRadar.x = ulWorld.x / (m_mapExtent.width() / RADAR_CELL_WIDTH);
- 	ulRadar.y = ulWorld.y / (m_mapExtent.height() / RADAR_CELL_HEIGHT);
+ 	ulRadar.x = ulWorld.x / (m_mapExtent.width() / static_cast<float>(RADAR_CELL_WIDTH));
+ 	ulRadar.y = ulWorld.y / (m_mapExtent.height() / static_cast<float>(RADAR_CELL_HEIGHT));
 
 	//
 	// convert radar point to actual pixel coords on the screen, shifted
@@ -377,7 +377,7 @@ void W3DRadar::drawSingleBeaconEvent( Int pixelX, Int pixelY, Int width, Int hei
 	Real maxEventSize = width / 10.0f;   // max size of the event marker
 	Int minEventSize = 6;     // min size of the event marker
 	Int eventSize;									 // current size of a marker to draw
-	const Real TIME_FROM_FULL_SIZE_TO_SMALL_SIZE = LOGICFRAMES_PER_SECOND * 1.5;
+	const Real TIME_FROM_FULL_SIZE_TO_SMALL_SIZE = static_cast<float>(LOGICFRAMES_PER_SECOND) * 1.5;
 	Real totalAnglesToSpin = 2.0f * PI;  ///< spin around this many angles going from big to small
 	UnsignedByte r, g, b, a;
 
@@ -476,7 +476,7 @@ void W3DRadar::drawSingleGenericEvent( Int pixelX, Int pixelY, Int width, Int he
 	Real maxEventSize = width / 2.0f;   // max size of the event marker
 	Int minEventSize = 6;     // min size of the event marker
 	Int eventSize;									 // current size of a marker to draw
-	const Real TIME_FROM_FULL_SIZE_TO_SMALL_SIZE = LOGICFRAMES_PER_SECOND * 1.5;
+	const Real TIME_FROM_FULL_SIZE_TO_SMALL_SIZE = static_cast<float>(LOGICFRAMES_PER_SECOND) * 1.5;
 	Real totalAnglesToSpin = 2.0f * PI;  ///< spin around this many angles going from big to small
 	UnsignedByte r, g, b, a;
 
@@ -708,8 +708,8 @@ void W3DRadar::renderObjectList( const RadarObject *listHead, TextureClass *text
 		const Coord3D *pos = obj->getPosition();
 
 		// compute object position as a radar blip
-		radarPoint.x = pos->x / (m_mapExtent.width() / RADAR_CELL_WIDTH);
-		radarPoint.y = pos->y / (m_mapExtent.height() / RADAR_CELL_HEIGHT);
+		radarPoint.x = pos->x / (m_mapExtent.width() / static_cast<float>(RADAR_CELL_WIDTH));
+		radarPoint.y = pos->y / (m_mapExtent.height() / static_cast<float>(RADAR_CELL_HEIGHT));
 
 		// get the color we're going to draw in
 		Color argbColor = rObj->getColor();
